requirements.txt
fitparse>=1.2.0
exif
matplotlib
Pillow
pyzbar
